requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.2.0
colorama==0.4.6
dataclasses-json==0.5.12
Django==4.2.3
django-rest-framework==0.1.0
djangorestframework==3.14.0
frozenlist==1.4.0
greenlet==2.0.2
idna==3.4
langchain==0.0.235
langsmith==0.0.7
marshmallow==3.19.0
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.4
numpy==1.25.1
openai==0.27.8
openapi-schema-pydantic==1.2.4
packaging==23.1
pydantic==1.10.11
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
SQLAlchemy==2.0.19
sqlparse==0.4.4
tenacity==8.2.2
tqdm==4.65.0
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.3
yarl==1.9.2
